requirements.txt
numpy
pydicom
PyQt5
matplotlib
opencv-python
mysql-connector-python
keras
django